requirements.txt
fastapi
uvicorn[standard]
google-adk
google-generativeai
pydantic
httpx
